// src/generated/models/Crc5b_codedirectoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crc5b_codedirectoriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crc5b_codedirectoriesstatecode = keyof typeof Crc5b_codedirectoriesstatecode;
export const Crc5b_codedirectoriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crc5b_codedirectoriesstatuscode = keyof typeof Crc5b_codedirectoriesstatuscode;

export interface Crc5b_codedirectoriesBase {
  crc5b_code: string;
  crc5b_codedirectoryid: string;
  crc5b_name: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crc5b_codedirectoriesstatecode;
  statuscode?: Crc5b_codedirectoriesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Crc5b_codedirectories extends Crc5b_codedirectoriesBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}