// src/generated/models/Crc5b_activitycodedirectoriesModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Crc5b_activitycodedirectoriesstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Crc5b_activitycodedirectoriesstatecode = keyof typeof Crc5b_activitycodedirectoriesstatecode;
export const Crc5b_activitycodedirectoriesstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Crc5b_activitycodedirectoriesstatuscode = keyof typeof Crc5b_activitycodedirectoriesstatuscode;

export interface Crc5b_activitycodedirectoriesBase {
  crc5b_activitycodedirectoryid: string;
  crc5b_cinnost: string;
  "crc5b_CodeDirectory@odata.bind": string;
  crc5b_kodcinnosti: string;
  importsequencenumber?: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Crc5b_activitycodedirectoriesstatecode;
  statuscode?: Crc5b_activitycodedirectoriesstatuscode;
  timezoneruleversionnumber?: string;
  utcconversiontimezonecode?: string;
}

export interface Crc5b_activitycodedirectories extends Crc5b_activitycodedirectoriesBase {
  crc5b_codedirectoryname?: string;
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: string;
  crc5b_codedirectory?: object;
  _crc5b_codedirectory_value?: string;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}